Fix CI compile error: totalOwnViews type mismatch (Int vs Long)
Article.viewsCount is Long, so sumOf{it.viewsCount} returns Long — the
previous ProfileUiState.totalOwnViews: Int field couldn't hold it.
Widened the field and ControlPanelSection's matching parameter to Long.

+++ android-native/app/src/main/kotlin/studio/ai/literium/literium_app/ui/screens/profile/ProfileScreen.kt
@@ -433,7 +433,7 @@ private fun SubTabPill(label: String, icon: androidx.compose.ui.graphics.vector.
  *  duplicating that already-built UI a second time inside this card. */
 @Composable
 private fun ControlPanelSection(
-    totalViews: Int,
+    totalViews: Long,
     articlesCount: Int,
     followersCount: Int,
     onOpenAds: () -> Unit,
